Update plugin REST API with new URLs and XML support.
The plugin REST interface is now under /api/v1.0/ and it also renders XML now as well as JSON (plugin update is still JSON only). To maximise code reuse, plugin structure is built as a tree of maps that can be rendered as JSON or XML. Same technique should work for other REST interfaces.