Use plugin releases to order RSS feed for latest plugins.
Fix for WEBSITE-8. Previously any change to a plugin's portal page would put the plugin at the top of the "latest plugins" feed, which isn't ideal. Now only plugin releases count in the ordering of the plugins. Also, the plugin summary is put into the feed in place of the contents of the description tab.