Use Joda Time for date domain class properties.
Working with the standard JDK and dates is too painful. Switching to Joda Time to improve the experience. Also, this allows for easy date formatting, for example as ISO for REST responses. The plugin list via JSON and XML now includes the 'last released' date as well as the licence information.