Fix plugin release synchronisation.
The synchronisation of plugin releases from the master plugin list wasn't working because the update code never got executed if the current release in the database matched the current release in the master plugin list. The synchronisation is now performed for every plugin all the time. This does make the script rather slow.

In addition, the query to check whether a plugin release already existed in the database was wrong since it only checked for a particular plugin, not the associated release version.