Fix caching problem with View Info wiki pages.
Wiki pages have a View Info link that displays a list of versions for that page and (depending on whether the user is logged in and what permissions he has) diff and rollback links. However, if a user clicked on the link and then logged in afterwards, he would still see the non-authenticated user view, i.e. no diff links. This was because the view was cached under a key that ignored the active user and his/her permissions. The cache key now incorporates that information, hence fixing the issue.