Fix exception in user registration form.
The app was doing some basic saving of form submission data before redirecting to the login page. Unfortunately it was a rather naive implementation that didn't work but did introduce a problem with duplicate fields when submitting from the login or registration pages. All that code has been removed because it doesn't work. At some point, it would be good to handle authentication during form submission better so users don't lose their data.