Force rollback on checked exceptions when publishing a plugin.
The database was getting into a confused state when the PluginUpdateService threw a checked exception because the plugin tabs were created, but the corresponding plugin entry was not. This then prevented users from publishing the plugin until the database was fixed. Checked exceptions now trigger a rollback so the database can't get into that state.