Allow plugin releases of up to 50Mb.
The size constraint on plugin zips for the PendingRelease domain class had been updated, but there was also a command object with the same set of constraints that hadn't been updated: PublishPluginCommand. The command object now imports the constraints from PendingRelease so we don't have to update them in two places.