Add event bus and use it for search indexing.
For multi-node deployments, the Lucene indexes need to be kept in sync. One of the easiest ways to do this is by doing event-based search indexing and making sure those events propagate to all nodes in a cluster. In this case, we're using the plugin platform event bus with a Spring Integration implementation that propagates those events via AMQP to all nodes. The Spring Integration configuration also ensures that a node's event bus does not see events that originating from that node, nor are events from another node propagated through the event bus back to the AMQP broker.