Use old cache mechanism for XML plugin list.
Springcache seemed to be having problems in production, generating an empty plugin list. Not sure what was going on, but now we simply generate the content as a string and dump it into an Ehcache instance. Hence the caching of the content is no longer tied to request processing, SiteMesh, etc. We can afford to do this because there is only ever one instance of the plugin list.