Fix exception in paginated search results.
The search results page would occasionally throw an exception saying that the property 'body' was missing on an object. It turns out that the pagination in the search results page was using the Searchable plugin's built-in results page rather than the application's override. The pagination links now send the requests to the correct URL - /search instead of /searchable.

Fixes WEBSITE-168.