Fix the downloads page to work with existing data.
Some refactoring of the data model around downloads broke the downloads page when applied to existing data. This is a further refactoring along with data migrations that better suits the existing data without losing the advantages of the new model. Basically, documentation is no longer treated as a separate top-level download but as a child distribution file of a standard Grails download. So now a download can have files of type 'binary', 'source' and 'documentation'.