Add initial data values for new columns.
The Download.lastRelease and Tutorial.lastUpdated fields both need initial
values for the app to work. At least one download needs to be the latest, so
the migration scripts now set 2.0.4 as latest. The 'lastUpdated' field seems
to cause problems for text indexing if it's not given a proper date.