Remove 'schemaName' from db migration pre-conditions.
The 'schemaName' property on certain preconditions meant that they only worked
for the given database. If you ran the app against a database with a different
name, those preconditions would fail.

Removing the 'schemaName' property makes the preconditions work on the current
database, solving the problem. Even better, this change does not affect the
checksums of the corresponding changesets, so Liquibase doesn't complain when
they are run against a database that has already had them applied. Yay!