Fix 'flash of content' problem on download page.
We use Modernizr to attach CSS classes to the <html> element for graceful
degradation of our pages. Unfortunately, Modernizr was being loaded at the
end of the page (because of the default disposition of 'defer' in Resources)
and that was resulting in the classes being added *after* page load. So you
saw the non-JS content just before Modernizr was loaded.

Very simple fix: Modernizr is now disposition 'head', so the CSS classes are
added at the beginning of page load.