Add LESS reloading via client-side less.js.
Server-side reloading of LESS is difficult in Grails when the @import statement
is used, so for development mode (run-app) we use client-side less.js rather
than the CSS generated from the LESS files on the server. When it comes to
generating the WAR file, we use the lesscss library for Java to compile the
LESS files to CSS, and it's the CSS files that are then referenced directly by
the pages.

This is a pretty ugly solution for a variety of reasons. To start with, we
shouldn't have to use client-side LESS, since Grails already has a reloading
mechanism that the Resources plugin hooks into. Unfortunately, there is no way
to know which LESS files need recompiling when one of them changes, because
you have to compile the top-level LESS file which ultimately imports the one
that you've changed.

On top of that, the lesscss Java library is pretty slow at compilation. During
WAR creation I have seen 6.5s and 19s. That's just not viable for automatic
reloading while you're developing. This is probably related to the library
using less.js on Rhino, but your guess is as good as mine.

Finally, I had to reorganise the LESS files so that they were no longer in
sub-directories. This was to work around an issue that client-side less.js
behaves differently with relative URLs and sub-directories compared to server-
side LESS. This GitHub pull request explains the general problem:

    https://github.com/cloudhead/less.js/pull/331

A faster LESS compiler for Java applications would go a long way to resolving
these problems. Any volunteers? :)