Fix news item and wiki page submissions.
CodeMirror and similar browser code editors typically attach themselves to a
<textarea> but leave that <textarea> empty. When the form is submitted, modern
browsers check the HTML 5 'required' attribute and block the submission if any
of those fields are empty. So you can't submit CodeMirror'ed fields if they
have the 'required' attribute.

This commit removes the 'required' attribute as it's the simplest, and possibly
only, fix.