Re-enable the publishing of plugins.
Some bits of the plugin controller were missing and had to be restored. I've
also taken the opportunity to provide an extra URL that allows clients to
track the status of a published plugin, since at the moment publishing often
looks like it's worked except the updated info doesn't appear in the portal.
This URL can be used by a new version of the Release plugin to let the user
know whether he or she needs to republish.

I have also made the deployment to the artifact repository more robust in the
event of connection time outs and the like. These are retried up to 3 times
before giving up.