Compile all top-level LESS files when creating WAR.
A new top-level LESS file, i.e. one that is included directly by a Resources
module, was added to the project, but it wasn't compiled to CSS when the WAR
was created. So, the CreateWarStart event handler now automatically compiles
all LESS files that don't begin with 'bootstrap-' or 'grails-'.