Fix redirect after linking accounts.
When a user linked their Twitter or Facebook account to an internal grails.org
one, he or she was redirected to a 404. This was because the Shiro OAuth
controller was not being passed a 'targetUri' parameter, and none was in the
session (this problem only occurred if the user clicked on the 'login' link).
We now set the 'targetUri' parameter to '/' if none is in the session.