Fix plugin publishing with big refactor.
There are several important changes here. The first is that PluginRelease now
supports the concept of a snapshot release. This allows us to handle requests
for 'latest.release' and 'latest.integration' separately and more in keeping
with their semantic meaning. It also ensures that the generated master plugin
list includes snapshot releases as well as normal ones.

The second big change is the extraction of much of PluginUpdateService into a
separate class (passing the same arguments between all the methods was just
getting silly). This has meant that we can correctly save snapshot as well as
normal releases with the correct download URL and other meta information.

Lastly, I've fixed the handling of PendingRelease instances so that they are
now deleted at the appropriate times. We should no longer end up with a large
number of stale PendingRelease records in the database due to failed plugin
publications.