Implement WEBSITE-198: add HTML shortening feature.
There is now a new class, HtmlShortener, that can be used to crop HTML content
to a maximum specified length. This doesn't crop based on the number of
characters in the markup, but on the number of displayed characters.

This class is used to shorten news items, tutorial descriptions, etc. after
the wiki text has been converted to HTML.