Implement WEBSITE-203: show extra plugin metadata.
The 'show plugin' page now shows extra metadata such as the licenses used by
a plugin and the organisation behind it. It will also now show all authors
associated with the plugin. This 'simple' change required some fairly serious
refactoring.

The database migration for supporting multiple authors only migrates the main
author for a plugin, i.e. the one specified by the 'author' and 'authorEmail'
fields, but when a plugin is published, any extra developers it defines will
be added to the database at that point. Then the multiple authors will be
refected in the UI.