Allow UserInfo without email address.
This is necessary so that we can create UserInfo instances for all plugin
authors, because some plugins only have an author name, not an author email.
In fact, we need a way for users to associate themselves with a plugin
without putting the email address in the plugin descriptor (which is often
publicly visible as many plugins are open source).