Load presentation in background.
The What is Grails? presentation can take a while to load on first show, which
is a problem because the first time a user sees it is the most important. The
Getting Started page was originally loading the presentation when the modal
opened, but this was too late.

It is now loaded as soon as the document is ready, i.e. in the background. One
thing to bear in mind with this approach is that the presentation auto-starts,
so when a user opens the modal, they could see the presentation half-way
through! To mitigate that, the presentation is hidden by an opaque pane and
the presentation is then reloaded. After about half a second, the pane fades
out and reveals the presentation from the start.