Fix WEBSITE-207: add titles to pages missing them.
A lot of GSP views were missing titles, which was causing problems with Disqus
(the same comment was appearing on multiple pages). Adding the titles should
hopefully fix WEBSITE-202 as well.

I've also removed a bunch of views that are no longer used and removed doctypes
from the ones I've modified. The doctype and content type settings are provided
by the layout.