Fix comments appearing on multiple pages.
The Disqus plugin has a bug in its <disqus:comments> tag that means it creates
the wrong URL for the 'disqus_url' Javascript variable. It was using the
requestURL property on the request, which of course gives the internal
.dispatch URL.

I have copied the plugin's taglib into the app for now and fixed it to use the
Grails link generator (because we need an absolute URL) with the forwardURI
property of the request.