Move less files back into top-level less directory.
There is a still a problem with less.js in that it is inconsistent in its
between the client and server sides. So if the less files are in sub-
directories, we see different url() paths in the css between run-app and
run-war.

We should revisit this once there is some sort of resolution on the following
issue:

    https://github.com/cloudhead/less.js/pull/961